feat: NSMutableAttributedString extension에 할인하는 상품의 할인가를 표시하고 정가에 취소선을 긋는 메서드 구현

+++ OpenMarket/OpenMarket/Model/DataType/Page.swift
@@ -31,9 +31,9 @@ struct Page: Decodable, Hashable {
     let description: String?
     let thumbnail: String
     let currency: Currency
-    let price: Double
-    let bargainPrice: Double
-    let discountedPrice: Double
+    let price: Int
+    let bargainPrice: Int
+    let discountedPrice: Int
     let stock: Int
     let createdAt: Date
     let issuedAt: Date
